+++ lib/src/material/modal_bottom_sheet_page.dart
@@ -1,3 +1,7 @@
+// Copyright (c) 2025. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 import 'package:flutter/material.dart';
 
 /// A declarative page-based modal bottom sheet that can be used with Navigator 2.0.